[BusyWorkers] refactor `init` and `add`
